Update README.md
As part of a personal project, I developed a website that provides comprehensive
information about the guitar store. Taking on a Front-End role, I utilized React and and
Tailwind CSS as the styling framework, to create a dynamic and interactive platform. The
website showcases various details about the guitar store, including its home, categories,
about, and help.